object texuring done

+++ Breakout/Pad.h
@@ -23,7 +23,7 @@ namespace Logic
 		~Pad();
 
 		float getRadius() { return radius; }
-		float getOrientation() { return rotation.z - PI / 2; }
+		float getOrientation() { return rotation.z - (float)PI / 2; }
 		float getPadRotation() { return rotation.z; }
 		float getWidth() { return width; }
