+++ Breakout/Physics.h
@@ -153,7 +153,7 @@ namespace Logic
 			
 			if(tBallPos.y - tRadius <= yIntersect)
 			{
-				Vec3 padRot = Vec3(std::cos(zrot + PI / 2), std::sin(zrot + PI /2), 0);
+				Vec3 padRot = Vec3((float)std::cos(zrot + PI / 2), (float)std::sin(zrot + PI /2), 0);
 				Vec3 newDir = planeReflection(_ball->getDirection(), padRot);
 				newDir.normalize();
 				_ball->setDirection(newDir.x, newDir.y, NULL);
